refactor: DDL 규격에 맞추어 모든 VO 클래스 데이터 타입 재정비

+++ src/main/java/com/example/smartmessaging/dto/vo/ChannelVO.java
@@ -13,9 +13,9 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @Builder
 public class ChannelVO extends BaseVO {
-    private BigDecimal id;
+    private Long id;
     private String channelType;
     private BigDecimal costPerMsg;
-    private BigDecimal maxLength;
+    private Integer maxLength;
     private Boolean isActive;
 }
