+++ src/main/java/com/example/smartmessaging/dto/vo/FailReasonVO.java
@@ -5,7 +5,6 @@ import lombok.Setter;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -13,8 +12,8 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @Builder
 public class FailReasonVO extends BaseVO {
-    private BigDecimal id;
+    private Long id;
     private String code;
     private String description;
-    private BigDecimal sendAttemptId;
+    private Long sendAttemptId;
 }
