+++ src/main/java/com/example/smartmessaging/dto/vo/SendAttemptVO.java
@@ -5,7 +5,6 @@ import lombok.Setter;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -13,10 +12,10 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @Builder
 public class SendAttemptVO extends BaseVO {
-    private BigDecimal id;
-    private BigDecimal sendTargetId;
-    private BigDecimal attemptOrder;
-    private BigDecimal channelId;
+    private Long id;
+    private Long sendTargetId;
+    private Integer attemptOrder;
+    private Long channelId;
     private Boolean isSucceeded;
     private String solapiMessageId;
 }
