+++ src/main/java/com/example/smartmessaging/dto/vo/TemplateVO.java
@@ -5,7 +5,6 @@ import lombok.Setter;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -13,14 +12,14 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @Builder
 public class TemplateVO extends BaseVO {
-    private BigDecimal id;
+    private Long id;
     private String title;
     private String content;
     private String kakaoTemplateCode;
     private String kakaoTemplateStatus;
     private Boolean isAiGenerated;
     private String category;
-    private BigDecimal cnt;
+    private Integer cnt;
     private String purpose;
-    private BigDecimal userId;
+    private Long userId;
 }
